feat(FN-1088): refactor how cypress tests interact with database
