+++ e2e-tests/portal/cypress/e2e/journeys/payment-report-officer/previous-reports/list-previous-reports.spec.js
@@ -12,6 +12,7 @@ context('List previous utilisation reports', () => {
 
   after(() => {
     cy.removeAllUtilisationReportDetails();
+    cy.closeDbConnection();
   });
 
   describe('On initial page load ', () => {
