+++ e2e-tests/portal/cypress/support/utils/closeDbConnection.js
@@ -1,0 +1,3 @@
+module.exports = () => {
+  cy.task('closeDbConnection');
+};
